temp commit

+++ Token.h
@@ -1,8 +1,8 @@
 #pragma once
 struct Token {
 	char* symbols;
-	int arguments = 0; // amount of operands
-	int index = 0;
+	short int arguments = 0; // amount of operands
+	short int index = 0;
 	size_t size = 0;
 	Token* next = nullptr;
 	Token* prev = nullptr;
